requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
datetime 